refactor(terminal): split terminal-service god-module
Extract cohesive leaf modules from the ~780 LOC terminal-service.ts (audit
architecture F5): terminal/font-size.ts, terminal/flow-control.ts,
terminal/path-utils.ts, terminal/pty-ready.ts. terminal-service.ts (now 551
LOC) keeps surface lifecycle, connectPty, setupListeners, and cwd polling,
and re-exports the moved symbols so the public API is unchanged. Pure
structural change; behavior identical.
